Added png graph of EV transmission vs. PWM value
This graph shows 1/3EV step measurements of light transmission reduction as a function of analog PWM value. Although it is labeled as analogWrite() values, it is actually the difference in analogWrite() values between the two pins driving the LCLV leads with an AC signal -- as implemented by this library with AN set to 1.